requirements.txt
discord.py
lark-parser
